metaespacio/requirements.txt
django < 3
django-crispy-forms
django-localflavor
django-tastypie
django-graphos
djangorestframework
django_extensions
django_debug_toolbar
pillow
gunicorn
django-oidc-provider
